Add files via upload
SO, I have made a few changes to the search
first being.. if there is only 1  of ' " ' in the query, then we would have a problem. I removed that problem by adding 2 or more  clause.   But odd number of double quotes will still create a problem. If you want I can take care of that but then the program will be lot slow.

Next, the way we search would add the phrase as is...and hence would have no effect. when we add a phrase to the query text... there has to be double quotes present so as the searcher would know that it is a phrase
That is done as well.